Assignment 42 completed

+++ app/controllers/comments_controller.rb
@@ -2,30 +2,55 @@ class CommentsController < ApplicationController
    before_action :require_sign_in
    before_action :authorize_user, only: [:destroy]
 
-   def create
-     @post = Post.find(params[:post_id])
-     comment = @post.comments.new(comment_params)
-     comment.user = current_user
+  def create
+    if @post = Post.find(params[:post_id])
+      comment = @post.comments.new(comment_params)
+      comment.user = current_user
 
-     if comment.save
-       flash[:notice] = "Comment saved successfully."
-       redirect_to [@post.topic, @post]
+      if comment.save
+        flash[:notice] = "Comment saved successfully."
+        redirect_to [@post.topic, @post]
+      else
+        flash[:alert] = "Comment failed to save."
+        redirect_to [@post.topic, @post]
+      end
+
+    else
+      @topic = Topic.find(params[:topic_id])
+      comment = @topic.comments.new(comment_params)
+      comment.user = current_user
+
+      if comment.save
+        flash[:notice] = "Comment saved successfully."
+        redirect_to [@topic]
+      else
+        flash[:alert] = "Comment failed to save."
+        redirect_to [@topic]
+      end
+    end
+  end
+
+  def destroy
+    comment = Comment.find(params[:id])
+
+    if @post = Post.find(params[:post_id])
+       if comment.destroy
+         flash[:notice] = "Comment was deleted."
+         redirect_to [@post.topic, @post]
+       else
+         flash[:alert] = "Comment couldn't be deleted. Try again."
+         redirect_to [@post.topic, @post]
+       end
      else
-       flash[:alert] = "Comment failed to save."
-       redirect_to [@post.topic, @post]
-     end
-   end
+       @topic = Topic.find(params[:topic_id])
 
-   def destroy
-     @post = Post.find(params[:post_id])
-     comment = @post.comments.find(params[:id])
-
-     if comment.destroy
-       flash[:notice] = "Comment was deleted."
-       redirect_to [@post.topic, @post]
-     else
-       flash[:alert] = "Comment couldn't be deleted. Try again."
-       redirect_to [@post.topic, @post]
+       if comment.destroy
+         flash[:notice] = "Comment was deleted."
+         redirect_to [@topic]
+       else
+         flash[:alert] = "Comment couldn't be deleted. Try again."
+         redirect_to [@topic]
+       end
      end
    end
